backend/requirements.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
openai>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pillow>=10.0.0
